WIP: try to extract metis helper module as class
