+++ polyphemus/lib/commands.rb
@@ -1,6 +1,7 @@
 require 'date'
 require 'logger'
 
+
 class Polyphemus
   class Help < Etna::Command
     usage 'List this help'
@@ -37,210 +38,9 @@ class Polyphemus
     end
   end
 
-  module WithMetisWaiverHelpers
-    include WithEtnaClients
-    include WithLogger
-
-    def project
-      raise "project must be implemented in subclasses!"
-    end
-
-    def patient_name_regex(patient_name)
-      /.*\/#{patient_name}-[\w-]+$/
-    end
-
-    def pool_regex(pool_name)
-      /.*\/#{pool_name}$/
-    end
-
-    def restrict_patient_data(patient_name)
-      # Move a folder from release_bucket to restrict_bucket
-      # If can't find the patient_name folder, throw exception.
-      # If we use paging, this would also have to fetch more pages.
-      # This expects a Patient name like MVIR1-HS10,
-      #   not an assay name, like MVIR1-HS10-D4BLD1-CYM2
-      patient_folders = release_folders.select { |folder|
-          folder.folder_path =~ patient_name_regex(patient_name)
-      }
-
-      # If we want to check by page, we would need to do
-      #   fetch more pages here. However, have to
-      #   make sure that we find the patient_name, so basically
-      #   have to fetch all folders...
-      msg = "No found folders in #{release_bucket_name} -- is #{patient_name} a valid patient name?"
-      logger.warn(msg) if patient_folders.length == 0
-
-      patient_folders.each { |folder|
-          rename_folder(
-              release_bucket_name,
-              folder.folder_path,
-              restrict_bucket_name,
-              folder.folder_path
-          )
-      }
-    end
-
-    def release_patient_data(patient_name)
-      # Move a folder from restrict_bucket to release_bucket
-      # If can't find the patient_name folder, throw exception.
-      # If we use paging, this would also have to fetch more pages.
-      # This expects a Patient name like MVIR1-HS10,
-      #   not an assay name, like MVIR1-HS10-D4BLD1-CYM2
-      patient_folders = restrict_folders.select { |folder|
-          folder.folder_path =~ patient_name_regex(patient_name)
-      }
-
-      # If we want to check by page, we would need to do
-      #   fetch more pages here. However, have to
-      #   make sure that we find the patient_name, so basically
-      #   have to fetch all folders...
-      msg = "No found folders in #{restrict_bucket_name} -- is #{patient_name} a valid patient name?"
-      logger.warn(msg) if patient_folders.length == 0
-
-      patient_folders.each { |folder|
-          rename_folder(
-              restrict_bucket_name,
-              folder.folder_path,
-              release_bucket_name,
-              folder.folder_path
-          )
-      }
-    end
-
-    def restrict_pool_data(pool_name)
-      # Move a folder from release_bucket to restrict_bucket
-      # Slightly different regex for pools because we should be able
-      #   to get the whole pool_name instead of a patient_name
-      #   that we can't match to folder names.
-      pool_folders = release_folders.select { |folder|
-          folder.folder_path =~ pool_regex(pool_name)
-      }
-
-      # If we want to check by page, we would need to do
-      #   fetch more pages here. However, have to
-      #   make sure that we find the patient_name, so basically
-      #   have to fetch all folders...
-      msg = "No found folders in #{release_bucket_name} -- is #{pool_name} a valid pool name?"
-      return logger.warn(msg) if pool_folders.length == 0
-
-      pool_folders.each { |folder|
-          rename_folder(
-              release_bucket_name,
-              folder.folder_path,
-              restrict_bucket_name,
-              folder.folder_path
-          )
-      }
-    end
-
-    def release_pool_data(pool_name)
-      # Move a folder from restrict_bucket to release_bucket
-      # Slightly different regex for pools because we should be able
-      #   to get the whole pool_name instead of a patient_name
-      #   that we can't match to folder names.
-      pool_folders = restrict_folders.select { |folder|
-          folder.folder_path =~ pool_regex(pool_name)
-      }
-
-      # If we want to check by page, we would need to do
-      #   fetch more pages here. However, have to
-      #   make sure that we find the patient_name, so basically
-      #   have to fetch all folders...
-      msg = "No found folders in #{restrict_bucket_name} -- is #{pool_name} a valid pool name?"
-      return logger.warn(msg) if pool_folders.length == 0
-
-      pool_folders.each { |folder|
-          rename_folder(
-              restrict_bucket_name,
-              folder.folder_path,
-              release_bucket_name,
-              folder.folder_path
-          )
-      }
-    end
-
-    private
-
-    def release_bucket_name
-      @release_bucket_name ||= Polyphemus.instance.config(:metis)[:release_bucket]
-    end
-
-    def restrict_bucket_name
-      @restrict_bucket_name ||= Polyphemus.instance.config(:metis)[:restrict_bucket]
-    end
-
-    def release_folders
-      @release_folders ||= fetch_folders(release_bucket_name).all
-    end
-
-    def restrict_folders
-      @restrict_folders ||= fetch_folders(restrict_bucket_name).all
-    end
-
-    def fetch_folders(bucket_name)
-      metis_client.list_all_folders(
-        Etna::Clients::Metis::ListFoldersRequest.new(
-          project_name: project, bucket_name: bucket_name)).folders
-    end
-
-    def parent_folder_path(folder_path)
-      folder_path.split('/')[0..-2].join('/')
-    end
-
-    def create_parent_folder(bucket_name, folder_path)
-      parent_path = parent_folder_path(folder_path)
-      logger.debug("Creating parent folder #{bucket_name}/#{folder_path}")
-      metis_client.create_folder(
-        Etna::Clients::Metis::CreateFolderRequest.new(
-          project_name: project,
-          bucket_name: bucket_name,
-          folder_path: parent_path
-      ))
-    end
-
-    def parent_exists?(bucket_name, folder_path)
-      # NOTE: this doesn't test if the folder_path itself exists
-      #   This can be confusing for root folders, because
-      #       they have no parents, so you don't need
-      #       to create anything.
-      parent_path = parent_folder_path(folder_path)
-      return true if parent_path.empty? # root folder
-
-      # returns 422 if the folder_path does not exist
-      begin
-          metis_client.list_folder(
-            Etna::Clients::Metis::ListFolderRequest.new(
-              project_name: project,
-              bucket_name: bucket_name,
-              folder_path: parent_path
-          ))
-      rescue Etna::Error => e
-          return false if e.status == 422
-          raise
-      end
-      return true
-    end
-
-    def rename_folder(source_bucket_name, source_folder_path, dest_bucket_name, dest_folder_path)
-      logger.debug("Renaming folder #{source_bucket_name}/#{source_folder_path} to #{dest_bucket_name}/#{dest_folder_path}")
-      create_parent_folder(dest_bucket_name, dest_folder_path) if !parent_exists?(dest_bucket_name, dest_folder_path)
-
-      logger.debug("Making the call to folder_rename route")
-      metis_client.rename_folder(
-        Etna::Clients::Metis::RenameFolderRequest.new(
-          bucket_name: source_bucket_name,
-          project_name: project,
-          folder_path: source_folder_path,
-          new_bucket_name: dest_bucket_name,
-          new_folder_path: dest_folder_path
-      ))
-    end
-  end
-
   class CascadeMvirPatientWaiverToRestricted < Etna::Command
     include WithEtnaClients
     include WithLogger
-    include WithMetisWaiverHelpers
 
     usage 'Updates any models whose restricted access does not match its relationship to patient'
 
@@ -293,7 +93,7 @@ class Polyphemus
         if all_restricted_pools.include? pool
           logger.info "Cytof pool #{pool} includes a restricted patient, restricting."
 
-          restrict_pool_data(pool)
+          metis_waiver_helper.restrict_pool_data(pool)
 
           update_request = Etna::Clients::Magma::UpdateRequest.new(project_name: project)
           update_request.update_revision('cytof_pool', pool, restricted: true)
@@ -301,7 +101,7 @@ class Polyphemus
         else
           logger.info "Cytof pool #{pool} does not include a restricted patient, relaxing."
 
-          release_pool_data(pool)
+          metis_waiver_helper.release_pool_data(pool)
 
           update_request = Etna::Clients::Magma::UpdateRequest.new(project_name: project)
           update_request.update_revision('cytof_pool', pool, restricted: false)
@@ -320,10 +120,10 @@ class Polyphemus
 
       if delete_metis_files
         # do metis movement attempt for now -- no auto-delete
-        restrict_patient_data(name)
+        metis_waiver_helper.restrict_patient_data(name)
       else
         # do metis movement attempt
-        restrict_patient_data(name)
+        metis_waiver_helper.restrict_patient_data(name)
       end
 
 
@@ -339,11 +139,16 @@ class Polyphemus
       # This code path should be --eventually consistent--  That is to say, we should ensure each operation
       # is idempotent (may need to be repeated), and that the patient is not marked restricted until all other
       # related tasks are complete and the state is consistent.
-      release_patient_data(name)
+      metis_waiver_helper.release_patient_data(name)
 
       update_request = Etna::Clients::Magma::UpdateRequest.new(project_name: project)
       update_request.update_revision('patient', name, restricted: false)
       magma_client.update(update_request)
+    end
+
+    def metis_waiver_helper
+      require_relative './metis/waiver_helper'
+      @metis_waiver_helper ||= MetisWaiverHelper.new(project)
     end
 
     def setup(config)
