+++ polyphemus/spec/examples.txt
@@ -1,16 +1,16 @@
 example_id                                    | status | run_time        |
 --------------------------------------------- | ------ | --------------- |
-./spec/cascade_mvir_restricted_spec.rb[1:1:1] | passed | 0.07517 seconds |
-./spec/cascade_mvir_restricted_spec.rb[1:1:2] | passed | 0.06381 seconds |
-./spec/cascade_mvir_restricted_spec.rb[1:2:1] | passed | 0.04603 seconds |
-./spec/cascade_mvir_restricted_spec.rb[1:2:2] | passed | 0.04327 seconds |
-./spec/metis_waiver_helper_spec.rb[1:1]       | passed | 0.00617 seconds |
-./spec/metis_waiver_helper_spec.rb[1:2]       | passed | 0.00352 seconds |
-./spec/metis_waiver_helper_spec.rb[1:3]       | passed | 0.02485 seconds |
-./spec/metis_waiver_helper_spec.rb[1:4]       | passed | 0.00431 seconds |
-./spec/metis_waiver_helper_spec.rb[1:5]       | passed | 0.0255 seconds  |
-./spec/metis_waiver_helper_spec.rb[1:6]       | passed | 0.00337 seconds |
-./spec/metis_waiver_helper_spec.rb[1:7]       | passed | 0.01387 seconds |
-./spec/metis_waiver_helper_spec.rb[1:8]       | passed | 0.00359 seconds |
-./spec/metis_waiver_helper_spec.rb[1:9]       | passed | 0.01605 seconds |
-./spec/metis_waiver_helper_spec.rb[1:10]      | passed | 0.00634 seconds |
+./spec/cascade_mvir_restricted_spec.rb[1:1:1] | failed | 0.01839 seconds |
+./spec/cascade_mvir_restricted_spec.rb[1:1:2] | failed | 0.00254 seconds |
+./spec/cascade_mvir_restricted_spec.rb[1:2:1] | failed | 0.00241 seconds |
+./spec/cascade_mvir_restricted_spec.rb[1:2:2] | failed | 0.0022 seconds  |
+./spec/metis_waiver_helper_spec.rb[1:1]       | failed | 0.00128 seconds |
+./spec/metis_waiver_helper_spec.rb[1:2]       | failed | 0.00111 seconds |
+./spec/metis_waiver_helper_spec.rb[1:3]       | failed | 0.0011 seconds  |
+./spec/metis_waiver_helper_spec.rb[1:4]       | failed | 0.00104 seconds |
+./spec/metis_waiver_helper_spec.rb[1:5]       | failed | 0.00106 seconds |
+./spec/metis_waiver_helper_spec.rb[1:6]       | failed | 0.00098 seconds |
+./spec/metis_waiver_helper_spec.rb[1:7]       | failed | 0.00102 seconds |
+./spec/metis_waiver_helper_spec.rb[1:8]       | failed | 0.00105 seconds |
+./spec/metis_waiver_helper_spec.rb[1:9]       | failed | 0.00137 seconds |
+./spec/metis_waiver_helper_spec.rb[1:10]      | failed | 0.01128 seconds |
